\name{models}
\alias{models}
\title{All Available Models}
\description{
  Returns list of model names.
}
\usage{
models()
}
\details{
  returns list of names (character strings) of models to be used
  as importance sampling distributions by \code{\link{bnmarg}}
  and \code{\link{bnlogl}}.
}
\value{
  A character vector.
}
\examples{
models()
}
\keyword{misc}
